+++ src/main/kotlin/com/github/com/cakevm/intellij_huff_plugin/language/psi/Utils.kt
@@ -14,6 +14,3 @@ inline fun <reified T : PsiElement> PsiElement.childOfType(strict: Boolean = tru
 
 val PsiElement.parentRelativeRange: TextRange
   get() = rangeRelativeTo(parent)
-
-val PsiElement.ancestors: Sequence<PsiElement>
-  get() = generateSequence(this) { it.parent }