+++ src/main/kotlin/com/github/com/cakevm/intellij_huff_plugin/language/psi/mixin/HuffMacroLabelMixin.kt
@@ -8,6 +8,7 @@ import com.intellij.ide.projectView.PresentationData
 import com.intellij.lang.ASTNode
 import com.intellij.psi.stubs.IStubElementType
 
+@Suppress("unused")
 abstract class HuffMacroLabelMixin : HuffStubbedNamedElementImpl<HuffMacroLabelStub>, HuffMacroLabel {
   constructor(node: ASTNode) : super(node)